requirement.txt
weblogo==3.7.12
flask==2.0.2
Pillow==9.5.0